+++ src/onegov.ballot/onegov/ballot/models/election/party_result.py
@@ -2,7 +2,6 @@ from onegov.core.orm import Base
 from onegov.core.orm.mixins import TimestampMixin
 from onegov.core.orm.types import UUID
 from sqlalchemy import Column
-from sqlalchemy import ForeignKey
 from sqlalchemy import Integer
 from sqlalchemy import Text
 from uuid import uuid4
@@ -16,8 +15,8 @@ class PartyResult(Base, TimestampMixin):
     #: identifies the party result
     id = Column(UUID, primary_key=True, default=uuid4)
 
-    #: the election this result belongs to
-    election_id = Column(Text, ForeignKey('elections.id'), nullable=False)
+    #: the election or election  compound this result belongs to
+    owner = Column(Text, nullable=False)
 
     # number of mandates
     number_of_mandates = Column(Integer, nullable=False, default=lambda: 0)